move extract_arguments into utils
